Changed integer types

+++ src/create.cpp
@@ -104,10 +104,10 @@ static void displayCharacterRaces() {
     clearToBottom(20);
     putString("Choose a race (? for Help):", 20, 2);
 
-    int y = 2;
-    int x = 21;
+    int8_t y = 2;
+    int8_t x = 21;
 
-    for (int i = 0; i < PLAYER_MAX_RACES; i++) {
+    for (uint8_t i = 0; i < PLAYER_MAX_RACES; i++) {
         char description[80];
 
         (void) sprintf(description, "%c) %s", i + 'a', character_races[i].name);
